+++ nodeconductor/structure/serializers.py
@@ -1157,13 +1157,14 @@ class BaseServiceProjectLinkSerializer(PermissionFieldFilteringMixin,
         lookup_field='uuid')
 
     service_name = serializers.ReadOnlyField(source='service.settings.name')
+    quotas = quotas_serializers.BasicQuotaSerializer(many=True, read_only=True)
 
     class Meta(object):
         model = NotImplemented
         fields = (
             'url',
             'project', 'project_name', 'project_uuid',
-            'service', 'service_uuid', 'service_name',
+            'service', 'service_uuid', 'service_name', 'quotas',
         )
         related_paths = ('project', 'service')
         extra_kwargs = {