+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ CHANGELOG
 
 Release NEXT
 ------------
+- Introduce ram, vcpu, storage quotas to service project link.
 - Drop SynchronizationStates, SynchronizationStateFilter, SynchronizableMixin.
 
 Release 0.134.0